+++ BankManagerConsoleApp/BankManagerConsoleApp/LinkedList/Node.swift
